Trying to get text selection to work.
I am starting a new branch to work on this.
I plan on making text nodes go into a buffer, then it will be easier to
traverse the nodes. If the node is removed, it will check to see if
it’s parent is only containing the cursor. If so it will remove that
parent and adjust the cursor position accordingly.