Fork: Information
The text-selection branch turned into more than just text-selection
including a complete rewrite of the application. So before it can be
merged back into the master branch I would at least like for text
formatting to be rewritten and included in the pull request. The branch
text-selection is now on hold until this branch is done.